requirements.txt
Django==1.8.3
django-registration-redux==1.2
Pillow==2.9.0
pycparser==2.14
wheel==0.24.0
